mini_project-main/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
numpy==1.26.4
opencv-python==4.9.0.80
tensorflow==2.16.1
Pillow==10.3.0
matplotlib==3.8.4
scikit-learn==1.4.2
werkzeug==3.0.3
requests==2.31.0
